+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.9] - 2015-04-14
+### Added
+ - Controllers as actions
+ - New internal flow
+
+### Changed
+ - Auto-create language strings
+
+
 ## [0.8.12] - 2015-04-10
 ### Added
  - New routing system